added react-hooks testing library and tests for useKeyPress hook
