+++ package.json
@@ -30,6 +30,7 @@
     "@testing-library/cypress": "8.0.2",
     "@testing-library/jest-dom": "5.16.4",
     "@testing-library/react": "13.2.0",
+    "@testing-library/react-hooks": "^8.0.1",
     "@testing-library/user-event": "14.2.0",
     "@types/jest": "27.5.1",
     "@types/node": "16.11.36",
